refactor(git): extract domain logic from git-write.ts handlers to services/git
- Move detectRepoOperationState, resolveGitDir to services/git/repoOperationState.ts, referencing OPERATION_SENTINEL_NAMES from gitRepoOperationState.ts
- Move parsePorcelainV2Conflicts, CONFLICT_LABELS to services/git/porcelainConflicts.ts
- Move scanStagedFilesForConflictMarkers to services/git/conflictMarkerScan.ts
- Add barrel index at services/git/index.ts
- Remove local StagingFileEntry, import from shared types
- Add tests for detectRepoOperationState (precedence, rebase progress)
- Update test imports to reference new service paths
